+++ core/src/com/astetyne/expirium/client/entity/SquirrelEntity.java
@@ -7,7 +7,7 @@ import com.badlogic.gdx.math.Vector2;
 
 public class SquirrelEntity extends Entity {
 
-    public SquirrelEntity(int id, Vector2 loc, PacketInputStream in) {
+    public SquirrelEntity(short id, Vector2 loc, PacketInputStream in) {
         super(EntityType.SQUIRREL, id, loc);
         setAnimator(new MoveableEntityAnimator(this, Res.SQUIRREL_IDLE, Res.SQUIRREL_IDLE));
     }
